refactor
changes regarding fetching data in init of viewmodel
